Fix/stop high precision waits (#641)
* default to Sleep for interval waiter, no spin phase

* keep sampling window placement precise even when interval timer misses

* change period range and loosen strictness

* name service work threads

* improving interval waiter cadence & telemetry poll setting

+++ IntelPresentMon/Core/source/kernel/Overlay.h
@@ -63,7 +63,7 @@ namespace p2c::kern
     private:
         // functions
         void AdjustOverlaySituation_(OverlaySpec::OverlayPosition position);
-        void UpdateGraphData_(double timestamp);
+        void UpdateGraphData_(uint64_t timestamp);
         void Render_();
         void UpdateCaptureStatusText_();
         void UpdateDataSets_();
